'Stack' multiple adapters
A series of adapters can be specified. For instance, first check in a
directory to see if an image has been uploaded; if not, try to download
it.

Each adapter does a node style `fn(err, success)` callback. `err`
queues up the next adapter.

TODO: The tiling shouldn't happen in the adapter. In the case of
multiple adapters (for instance 'local directory, mediabin') shouldn't
the adapter only be responsible for grabbing an image? Then the tiling
should happen after the image location has been determined.

Unless maybe there's an adapter that finds pre-tiled images…